+++ tests/mock_node.h
@@ -56,24 +56,24 @@ static const char *Token(const void *node) {
   return ((Node *)node)->token.data();
 }
 
-static int ChildrenSize(const void *node) {
-  return static_cast<int>(((Node *)node)->children.size());
+static size_t ChildrenSize(const void *node) {
+  return ((Node *)node)->children.size();
 }
 
 static void *ChildAt(const void *node, int index) {
   return ((Node *)node)->children.at(index);
 }
 
-static int RolesSize(const void *node) {
-  return static_cast<int>(((Node *)node)->roles.size());
+static size_t RolesSize(const void *node) {
+  return ((Node *)node)->roles.size();
 }
 
 static uint16_t RoleAt(const void *node, int index) {
   return ((Node *)node)->roles.at(index);
 }
 
-static int PropertiesSize(const void *node) {
-  return static_cast<int>(((Node *)node)->properties.size());
+static size_t PropertiesSize(const void *node) {
+  return ((Node *)node)->properties.size();
 }
 
 static const char *PropertyKeyAt(const void *node, int index) {
